+++ assets/js/dashboard/util/url-search-params.ts
@@ -1,3 +1,7 @@
+import {
+  getSearchToSetSegmentFilter,
+  SavedSegment
+} from '../filtering/segments'
 import { Filter, FilterClauseLabels } from '../query'
 import { v1 } from './url-search-params-v1'
 import { v2 } from './url-search-params-v2'
@@ -230,8 +234,10 @@ function isAlreadyRedirected(searchParams: URLSearchParams) {
   The purpose of this function is to redirect users from one of the previous versions to the current version, 
   so previous dashboard links still work.
 */
-export function getRedirectTarget(windowLocation: Location): null | string {
-  const searchParams = new URLSearchParams(windowLocation.search)
+export function maybeGetLatestReadableSearch(
+  searchString: string
+): null | string {
+  const searchParams = new URLSearchParams(searchString)
   if (isAlreadyRedirected(searchParams)) {
     return null
   }
@@ -242,27 +248,67 @@ export function getRedirectTarget(windowLocation: Location): null | string {
 
   const isV2 = v2.isV2(searchParams)
   if (isV2) {
-    return `${windowLocation.pathname}${stringifySearch({ ...v2.parseSearch(windowLocation.search), [REDIRECTED_SEARCH_PARAM_NAME]: 'v2' })}`
+    return stringifySearch({
+      ...v2.parseSearch(searchString),
+      [REDIRECTED_SEARCH_PARAM_NAME]: 'v2'
+    })
   }
 
-  const searchRecord = v2.parseSearch(windowLocation.search)
+  const searchRecord = v2.parseSearch(searchString)
   const isV1 = v1.isV1(searchRecord)
 
   if (!isV1) {
     return null
   }
 
-  return `${windowLocation.pathname}${stringifySearch({ ...v1.parseSearchRecord(searchRecord), [REDIRECTED_SEARCH_PARAM_NAME]: 'v1' })}`
+  return stringifySearch({
+    ...v1.parseSearchRecord(searchRecord),
+    [REDIRECTED_SEARCH_PARAM_NAME]: 'v1'
+  })
+}
+
+/**
+ * It's possible to set a particular segment to be always applied on the data on dashboards accessed with a shared link.
+ * This function ensures that the particular segment filter is set to the URL string on initial page load.
+ * Other functions ensure that it can't be removed.
+ */
+export function getSearchWithEnforcedSegment(
+  searchString: string,
+  enforcedSegment: Pick<SavedSegment, 'id' | 'name'>
+): string {
+  const searchRecord = parseSearch(searchString)
+  return stringifySearch(
+    getSearchToSetSegmentFilter(enforcedSegment)(searchRecord)
+  )
 }
 
 /** Called once before React app mounts. If legacy url search params are present, does a redirect to new format. */
-export function redirectForLegacyParams(
+export function maybeDoFERedirect(
   windowLocation: Location,
-  windowHistory: History
+  windowHistory: History,
+  enforcedSegment: Pick<SavedSegment, 'id' | 'name'> | null
 ) {
-  const redirectTargetURL = getRedirectTarget(windowLocation)
-  if (redirectTargetURL === null) {
+  const originalSearchString = windowLocation.search
+
+  let updatedSearchString = maybeGetLatestReadableSearch(originalSearchString)
+
+  if (enforcedSegment) {
+    updatedSearchString = getSearchWithEnforcedSegment(
+      updatedSearchString ?? originalSearchString,
+      enforcedSegment
+    )
+  }
+
+  if (
+    updatedSearchString === null ||
+    updatedSearchString === originalSearchString
+  ) {
     return
   }
-  windowHistory.pushState({}, '', redirectTargetURL)
+
+  windowHistory.pushState(
+    {},
+    '',
+    `${windowLocation.pathname}${updatedSearchString}`
+  )
 }